requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.117.1
h11==0.16.0
idna==3.10
pydantic==2.11.9
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
SQLAlchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.2   # optional, for migrations
psycopg2-binary==2.9.9  # if you want sync
passlib[bcrypt]==1.7.4  # password hashing
bcrypt==4.0.1
python-jose[cryptography]==3.3.0  # JWT support
python-multipart
